feat(evolution): 組み合わせ PoC Combo-A/C (llive synergy)
単体 PoC を合成し、ブレストが指摘した組み合わせ効果を定量化。

- Combo-A (Antifragile x Speculative Mesh): panic 探索候補を idle peer へ並行投機 →
  脱出率 100% を保ちつつ 1 世代の wall-clock を W 分の 1 に。W=2/4/8 で 1.82x/3.41x/6.82x。
  W=1 は mesh overhead で 0.91x (honest: 並列度なしは純損)。
- Combo-C (Antifragile x 予測検証ゲート): panic は無効 ChangeOp burst を生む (panic 47.8/60
  世代)。前段ゲートが高 invalid burst で 80% 削減 → run 全体で verifier コスト 69% 削減
  (通常区間のみなら 25%)。antifragile が gate の価値を増幅。
- tests/unit/test_combo_bench.py: 4 ケース。ruff + mypy strict green。

goal: 全PoC 組み合わせ Combo-A/C

+++ src/llive/evolution/combo_bench.py
@@ -95,8 +95,8 @@ def combo_c(
     n_gen = results[0].n_gen
     mean_normal = max(0.0, n_gen - mean_panic)
 
-    panic_ops = max(1, int(round(mean_panic * ops_per_gen)))
-    normal_ops = max(1, int(round(mean_normal * ops_per_gen)))
+    panic_ops = max(1, round(mean_panic * ops_per_gen))
+    normal_ops = max(1, round(mean_normal * ops_per_gen))
 
     g_panic = simulate(GateModel(invalid_rate=panic_invalid_rate, n_candidates=panic_ops), seed=0)
     g_normal = simulate(GateModel(invalid_rate=normal_invalid_rate, n_candidates=normal_ops), seed=0)
